just use TypeToken.of(BrooklynJacksonType)
back out the RegisteredTypeToken and the old TypeTokenOrRegisteredType,
because the BrooklynJacksonType (extending JavaType implementing Type)
works as does TypeToken.of(BrooklynJacksonType).
