+++ core/src/test/java/org/apache/brooklyn/core/resolve/jackson/BrooklynJacksonTypeTest.java
@@ -18,23 +18,14 @@
  */
 package org.apache.brooklyn.core.resolve.jackson;
 
-import com.fasterxml.jackson.databind.JavaType;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.google.common.collect.Iterables;
 import com.google.common.reflect.TypeToken;
-import java.lang.reflect.Type;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.function.Function;
-import org.apache.brooklyn.core.resolve.jackson.BeanWithTypeUtils.RegisteredTypeToken;
-import org.apache.brooklyn.core.resolve.jackson.BrooklynRegisteredTypeJacksonSerialization.BrooklynJacksonType;
 import org.apache.brooklyn.core.typereg.BasicTypeImplementationPlan;
 import org.apache.brooklyn.core.typereg.RegisteredTypes;
 import org.apache.brooklyn.test.Asserts;
 import org.apache.brooklyn.util.core.flags.BrooklynTypeNameResolution;
 import org.apache.brooklyn.util.guava.Maybe;
 import org.apache.brooklyn.util.guava.TypeTokens;
-import org.apache.brooklyn.util.time.Duration;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
@@ -61,7 +52,7 @@ public class BrooklynJacksonTypeTest {
     public static TypeToken<?> parseTestTypes(String s1) {
         return BrooklynTypeNameResolution.parseTypeToken(s1, s2 -> {
             if ("foo".equals(s2))
-                return Maybe.of((TypeToken<?>) RegisteredTypeToken.of(RegisteredTypes.bean("foo", "1", new BasicTypeImplementationPlan("x", null))));
+                return Maybe.of(BrooklynJacksonType.asTypeToken(RegisteredTypes.bean("foo", "1", new BasicTypeImplementationPlan("x", null))));
             if ("iterable".equals(s2)) return Maybe.of(TypeToken.of(Iterable.class));
             return BrooklynTypeNameResolution.getTypeTokenForBuiltInTypeName(s2);
         }).get();
