+++ core/src/test/java/org/apache/brooklyn/util/core/flags/BrooklynTypeNameResolutionTest.java
@@ -22,16 +22,8 @@ import com.google.common.collect.Iterables;
 import com.google.common.reflect.TypeToken;
 import java.lang.reflect.ParameterizedType;
 import java.util.List;
-import java.util.function.Function;
-import org.apache.brooklyn.api.typereg.RegisteredType;
-import org.apache.brooklyn.core.resolve.jackson.BeanWithTypeUtils;
-import org.apache.brooklyn.core.resolve.jackson.BeanWithTypeUtils.RegisteredTypeToken;
 import org.apache.brooklyn.core.resolve.jackson.BrooklynJacksonTypeTest;
-import org.apache.brooklyn.core.typereg.BasicRegisteredType;
-import org.apache.brooklyn.core.typereg.BasicTypeImplementationPlan;
-import org.apache.brooklyn.core.typereg.RegisteredTypes;
 import static org.apache.brooklyn.util.core.flags.BrooklynTypeNameResolution.parseTypeGenerics;
-import org.apache.brooklyn.util.guava.Maybe;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.testng.Assert;
